+++ src/App.tsx
@@ -1,23 +1,21 @@
-import React from 'react';
-import './App.css';
-import Dashboard from './Dashboard'
-import { BrowserRouter } from 'react-router-dom';
-import { RecoilRoot } from 'recoil';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { grey } from '@mui/material/colors';
-import {
-  QueryClient,
-  QueryClientProvider,
-} from 'react-query'
+import React from "react";
+import "./App.css";
+import Dashboard from "./Dashboard";
+import { BrowserRouter } from "react-router-dom";
+import { RecoilRoot } from "recoil";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { grey } from "@mui/material/colors";
+import { QueryClient, QueryClientProvider } from "react-query";
+import ErrorBoundary from "./components/ErrorBoundary";
 
 const theme = createTheme({
   components: {
     MuiAppBar: {
       styleOverrides: {
         colorPrimary: {
-          backgroundColor: 'white',
+          backgroundColor: "white",
           color: grey[900],
-          boxShadow: "0px 1px 1px -1px"
+          boxShadow: "0px 1px 1px -1px",
         },
         // root: ({ ownerState }) => ({
         //   ...(ownerState.color === 'primary' && {
@@ -26,8 +24,8 @@ const theme = createTheme({
         //     }),
         // }),
       },
-    }
-  }
+    },
+  },
 });
 
 const queryClient = new QueryClient();
@@ -37,7 +35,9 @@ function App() {
       <QueryClientProvider client={queryClient}>
         <ThemeProvider theme={theme}>
           <BrowserRouter>
-            <Dashboard />
+            <ErrorBoundary>
+              <Dashboard />
+            </ErrorBoundary>
           </BrowserRouter>
         </ThemeProvider>
       </QueryClientProvider>